+++ src/jdbc/JdbcPaciente.java
@@ -49,12 +49,11 @@ public class JdbcPaciente {
                     preparedStmt.setString(4, paciente.getObservaciones());
                     preparedStmt.setInt(5, paciente.getHabitacion());
                     do {
-                        System.out.println("¿Seguro que quiere insertar este paciente? 1.-Sí 2.-No: \n" + paciente.toString());
+                        System.out.println("¿Seguro que quiere insertar este paciente? 1.-Sí 2.-No: ");
                         int insertar = sc.nextInt();
                         switch (insertar) {
                             case 1 -> {
-                                preparedStmt.execute(); //Se insertan los datos
-                                //Se muestra lo que se ha introducido
+                                preparedStmt.execute();
                                 System.out.println("PACIENTE INTRODUCIDO");
                             }
                             case 2 -> MenuMedico.menuMedico(conn);
@@ -80,7 +79,6 @@ public class JdbcPaciente {
         }
     }
     public static void borrarPaciente(Connection conn) {
-
         try {
             Scanner sc = new Scanner(System.in);
             System.out.println("""
@@ -94,25 +92,31 @@ public class JdbcPaciente {
                     Paciente.mostrarPacientes(consultarPaciente(conn));
                     System.out.println("Introducir NHC del paciente que desea borrar: ");
                     int nhc = sc.nextInt();
-                    do {
-                        System.out.println("¿Seguro que quiere borrar este registro? 1.-Sí 2.-No");
-                        int borrar = sc.nextInt();
-                        switch (borrar) {
-                            case 1 -> {
-                                String query = "DELETE FROM paciente WHERE nhc = ?";
-                                PreparedStatement preparedStmt = conn.prepareStatement(query);
-                                preparedStmt.setInt(1, nhc);
+                    if (consultarSiExiste(conn,nhc) != 0){
+                        do {
+                            System.out.println("¿Seguro que quiere borrar este registro? 1.-Sí 2.-No");
+                            int borrar = sc.nextInt();
+                            switch (borrar) {
+                                case 1 -> {
+                                    String query = "DELETE FROM paciente WHERE nhc = ?";
+                                    PreparedStatement preparedStmt = conn.prepareStatement(query);
+                                    preparedStmt.setInt(1, nhc);
 
-                                preparedStmt.execute();
-                                System.out.println("PACIENTE BORRADO");
+                                    preparedStmt.execute();
+                                    System.out.println("PACIENTE BORRADO");
+                                }
+                                case 2 -> borrarPaciente(conn);
+                                default -> {
+                                    System.out.println("Número fuera de rango. Vuelva a intoducirlo");
+                                    borrado = false;
+                                }
                             }
-                            case 2 -> borrarPaciente(conn);
-                            default -> {
-                                System.out.println("Número fuera de rango. Vuelva a intoducirlo");
-                                borrado = false;
-                            }
-                        }
-                    } while (!borrado);
+                        } while (!borrado);
+
+                    } else {
+                        System.out.println("El paciente no existe");
+                        borrarPaciente(conn);
+                    }
                 }
                 case 2 -> MenuMedico.menuMedico(conn);
                 default -> {
@@ -127,8 +131,8 @@ public class JdbcPaciente {
             System.out.println("Error indeterminado");
             borrarPaciente(conn);
         }
-    }
-    public static void editarPaciente(Connection conn) throws SQLException {
+    } //
+    public static void editarPaciente(Connection conn) throws SQLException { //TODO: Controlar que el nhc introducido realmente existe
         String query = "UPDATE paciente SET ";
         boolean seguir = true;
         boolean edicion = true;
@@ -145,81 +149,86 @@ public class JdbcPaciente {
                 case 1 -> {
                     Paciente.mostrarPacientes(JdbcPaciente.consultarPaciente(conn));
                     System.out.println("Introducir NHC del paciente que desea editar: ");
-                    int codigo = scInt.nextInt();
-                    System.out.println("""
+                    int nhc = scInt.nextInt();
+                    if (consultarSiExiste(conn,nhc) != 0){
+                        System.out.println("""
                             ¿Qué campo quiere editar del paciente?:\s
                             1.-Nombre\s
                             2.-Apellidos\s
                             3.-Observaciones\s
                             4.-Habitación\s
                             5.-Volver a Menú Médico""");
-                    int campo = scInt.nextInt();
-                    switch (campo) {
-                        case 1 -> {
-                            System.out.println("Introduzca el nuevo nombre: ");
-                            String nombre = scString.nextLine();
-                            query += "nombre = ? WHERE nhc= ?";
-                            preparedStmt = conn.prepareStatement(query);
-                            preparedStmt.setString(1, nombre);
-                            preparedStmt.setInt(2, codigo);
-                        }
-                        case 2 -> {
-                            System.out.println("Introduzca los nuevos apellidos: ");
-                            String apellidos = scString.nextLine();
-                            query += "apellidos = ? WHERE nhc = ?";
-                            preparedStmt = conn.prepareStatement(query);
-                            preparedStmt.setString(1, apellidos);
-                            preparedStmt.setInt(2, codigo);
-                        }
-                        case 3 -> {
-                            System.out.println("Introduzca las nuevas observaciones: ");
-                            String observaciones = scString.nextLine();
-                            query += "observaciones = ? WHERE nhc = ?";
-                            preparedStmt = conn.prepareStatement(query);
-                            preparedStmt.setString(1, observaciones);
-                            preparedStmt.setInt(2, codigo);
-                        }
-                        case 4 -> {
-                            System.out.println("Introduzca la nueva habitación: ");
-                            String habitacion = scString.nextLine();
-                            query += "habitacion = ? WHERE nhc = ?";
-                            preparedStmt = conn.prepareStatement(query);
-                            preparedStmt.setString(1, habitacion);
-                            preparedStmt.setInt(2, codigo);
-                        }
-                        case 5 -> MenuMedico.menuMedico(conn);
-                        default -> {
-                            System.out.println("Número fuera de rango. vuelva a introducir un número: ");
-                            editarPaciente(conn);
-                        }
-                    }
-                    do {
-                        System.out.println("¿Seguro que quiere editar este paciente? 1.-Sí 2.-No");
-                        int editar = scInt.nextInt();
-                        switch (editar) {
+                        int campo = scInt.nextInt();
+                        switch (campo) {
                             case 1 -> {
-                                preparedStmt.execute();
-                                System.out.println("PACIENTE EDITADO");
+                                System.out.println("Introduzca el nuevo nombre: ");
+                                String nombre = scString.nextLine();
+                                query += "nombre = ? WHERE nhc= ?";
+                                preparedStmt = conn.prepareStatement(query);
+                                preparedStmt.setString(1, nombre);
+                                preparedStmt.setInt(2, nhc);
                             }
-                            case 2 -> MenuMedico.menuMedico(conn);
+                            case 2 -> {
+                                System.out.println("Introduzca los nuevos apellidos: ");
+                                String apellidos = scString.nextLine();
+                                query += "apellidos = ? WHERE nhc = ?";
+                                preparedStmt = conn.prepareStatement(query);
+                                preparedStmt.setString(1, apellidos);
+                                preparedStmt.setInt(2, nhc);
+                            }
+                            case 3 -> {
+                                System.out.println("Introduzca las nuevas observaciones: ");
+                                String observaciones = scString.nextLine();
+                                query += "observaciones = ? WHERE nhc = ?";
+                                preparedStmt = conn.prepareStatement(query);
+                                preparedStmt.setString(1, observaciones);
+                                preparedStmt.setInt(2, nhc);
+                            }
+                            case 4 -> {
+                                System.out.println("Introduzca la nueva habitación: ");
+                                String habitacion = scString.nextLine();
+                                query += "habitacion = ? WHERE nhc = ?";
+                                preparedStmt = conn.prepareStatement(query);
+                                preparedStmt.setString(1, habitacion);
+                                preparedStmt.setInt(2, nhc);
+                            }
+                            case 5 -> MenuMedico.menuMedico(conn);
                             default -> {
-                                System.out.println("Número fuera de rango. Vuelva a intoducirlo");
-                                edicion = false;
+                                System.out.println("Número fuera de rango. vuelva a introducir un número: ");
+                                editarPaciente(conn);
                             }
                         }
-                    } while (!edicion);
-                    do {
-                        System.out.println("¿Quiere editar otro campo? 1.-Sí 2.-Volver a Menú Médico");
-                        int continuar = scInt.nextInt();
-                        switch (continuar) {
-                            case 1 -> editarPaciente(conn);
-                            case 2 -> MenuMedico.menuMedico(conn);
-                            default -> {
-                                System.out.println("Número fuera de rango. Vuelva a intoducirlo");
-                                seguir = false;
+                        do {
+                            System.out.println("¿Seguro que quiere editar este paciente? 1.-Sí 2.-No");
+                            int editar = scInt.nextInt();
+                            switch (editar) {
+                                case 1 -> {
+                                    preparedStmt.execute();
+                                    System.out.println("PACIENTE EDITADO");
+                                }
+                                case 2 -> MenuMedico.menuMedico(conn);
+                                default -> {
+                                    System.out.println("Número fuera de rango. Vuelva a intoducirlo");
+                                    edicion = false;
+                                }
                             }
-                        }
-                    } while (!seguir);
+                        } while (!edicion);
+                        do {
+                            System.out.println("¿Quiere editar otro campo? 1.-Sí 2.-Volver a Menú Médico");
+                            int continuar = scInt.nextInt();
+                            switch (continuar) {
+                                case 1 -> editarPaciente(conn);
+                                case 2 -> MenuMedico.menuMedico(conn);
+                                default -> {
+                                    System.out.println("Número fuera de rango. Vuelva a intoducirlo");
+                                    seguir = false;
+                                }
+                            }
+                        } while (!seguir);
+                    } else {
+                        System.out.println("El paciente no existe");
+                        editarPaciente(conn);
+                    }
                 }
                 case 2 -> MenuMedico.menuMedico(conn);
                 default -> {
@@ -234,5 +243,16 @@ public class JdbcPaciente {
             System.out.println("Error indeterminado");
             editarPaciente(conn);
         }
+    }
+
+    private static int consultarSiExiste(Connection conn, int nhc) throws SQLException {
+        int resultado = 0;
+        String query = "SELECT EXISTS (SELECT * FROM paciente WHERE nhc=" + nhc + ") AS resultado";
+        Statement stmt = conn.createStatement();
+        ResultSet rs = stmt.executeQuery(query);
+        while (rs.next()){
+            resultado = rs.getInt("resultado");
+        }
+        return resultado;
     }
 }
